fix faiss tests not running
