+++ conftest.py
@@ -2,7 +2,7 @@ def pytest_addoption(parser):
     parser.addoption(
         "--document_store_type",
         action="store",
-        default="elasticsearch, faiss, sql, memory, milvus1, milvus, weaviate, pinecone",
+        default="elasticsearch, faiss, sql, memory, milvus1, milvus, weaviate, pinecone, opensearch",
     )
